ci: change verifier configuration to use system property
This commit changes how plugin verifier is run from the Github action. Specifically,the `-P` option in the Gradle command used to run the plugin verifier task is replaces with a `-D` option to set the plugin verifier home directory as a system property. Additionally, permissions are added to allow write access to contents.