build: improve setting environment variables for Qodana and Publication
Changed the `build.gradle.kts` file in order to use environment variables instead of system calls to set properties for Qodana and Publication. New function `environment()` was added to the file to retrieve environment variables and mapped them accordingly in the tasks section of the file. Qodana task now uses the `environment()` function to get the value of `showReport` while Publication task uses environment variables to set `token`.