ci: update GitHub workflow to use Gradle build action
The .github/workflows/run-ui-tests.yml file has been adjusted to use the gradle/gradle-build-action instead of caching gradle. This change is made to improve the setup process of Gradle in the workflow. With this update, the setup will be handled by the Gradle build action, ensuring a more accurate and efficient setup.