docs: update README.md to enhance plugin description
In this commit, the description of the Tauri plugin in the README.md has been updated. The aim was to give a more comprehensive overview of the plugin's functionality, including how it simplifies the process of setting up a new project with support for various front-end frameworks. The instructions for plugin installation have also been rephrased.