build(deps-dev): update Gradle version to 8.5
The project's Gradle version is upgraded from 8.2.1 to 8.5. This improvement includes updates on gradle-wrapper.properties and gradle.properties files, and also brings a modification in the gradlew script. ShellCheck warnings were addressed and more details about the argument collection for the java command have been added to the gradlew script.